fix: fetch all currencies fixed
The commit fixes the issue with fetching all currencies. It updates the `.gitignore` file to include `package-lock.json` as a dependency. Additionally, it adds `@types/cheerio` as a dependency in the `package.json` file. In the `currency.service.ts` file, it modifies the regular expression used for filtering currency code elements and improves the iteration logic to push each currency code into an array.